// frontend/src/app/models/category.enum.ts
export enum Category {
  PERFORMANCE = "Performance",
  TECHNICAL_SKILLSET = "Technical Skillset",
  TECHNICAL_PREDISPOSITIONS = "Technical Predispositions",
  SALES = "Sales",
  RECRUITING = "Recruiting",
  TEAMWORK = "Teamwork",
  COACHING = "Coaching",
  PRODYNA_INSIGHTS = "Prodyna Insights",
  OVERALL = "Overall"
}